+++ module/item/entity.js
@@ -352,7 +352,6 @@ export class ItemPF extends Item {
   static get defaultContextNote() {
     return {
       text: "",
-      target: "",
       subTarget: "",
     };
   }
@@ -880,18 +879,6 @@ export class ItemPF extends Item {
 
             delete data[entry[0]];
           });
-
-          // Special case for context notes
-          if (kArr.key === "data.contextNotes") {
-            for (let obj of arr) {
-              if (obj.target) {
-                const subTargets = Object.keys(this.getContextNoteSubTargets(obj.target));
-                if (!subTargets.includes(obj.subTarget)) {
-                  obj.subTarget = subTargets[0];
-                }
-              }
-            }
-          }
 
           linkData(srcData, data, kArr.key, arr);
         }
@@ -4057,44 +4044,6 @@ export class ItemPF extends Item {
         result = { ...result, crit: "PF1.CritDamageBonusFormula", nonCrit: "PF1.NonCritDamageBonusFormula" };
       }
     }
-    return result;
-  }
-
-  /**
-   * Generates lists of context note subtargets this item can have.
-   *
-   * @param {string} target - The target key, as defined in CONFIG.PF1.buffTargets.
-   * @returns {object.<string, string>} A list of changes
-   */
-  getContextNoteSubTargets(target) {
-    let result = {};
-    // Add specific skills
-    if (target === "skill") {
-      if (this.parentActor == null) {
-        for (let [s, skl] of Object.entries(CONFIG.PF1.skills)) {
-          result[`skill.${s}`] = skl;
-        }
-      } else {
-        const actorSkills = mergeObject(duplicate(CONFIG.PF1.skills), this.parentActor.data.data.skills);
-        for (let [s, skl] of Object.entries(actorSkills)) {
-          if (!skl.subSkills) {
-            if (skl.custom) result[`skill.${s}`] = skl.name;
-            else result[`skill.${s}`] = CONFIG.PF1.skills[s];
-          } else {
-            for (let [s2, skl2] of Object.entries(skl.subSkills)) {
-              result[`skill.${s}.subSkills.${s2}`] = `${CONFIG.PF1.skills[s]} (${skl2.name})`;
-            }
-          }
-        }
-      }
-    }
-    // Add static subtargets
-    else if (hasProperty(CONFIG.PF1.contextNoteTargets, target)) {
-      for (let [k, v] of Object.entries(CONFIG.PF1.contextNoteTargets[target])) {
-        if (!k.startsWith("_") && !k.startsWith("~")) result[k] = v;
-      }
-    }
-
     return result;
   }