+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Loot sheets now add coinage to the total sell value at the bottom
 - Added a world setting for an alternative reach rule, which hides the outer corners of 10-ft reach measurements
 - Added a new method of automatic spell slot calculation (thanks claudekennilol!)
+- Added a new way of choosing change and context note targets
 
 ### API